// src/booking/types/booking.hotel.response.ts
export interface BookingHotelResponse {
  searchQueries: SearchQueries;
}

export interface SearchQueries {
  search: Search;
}

export interface Search {
  destinationLocation: DestinationLocation;
  searchMeta: SearchMeta;
  filters: Filter[];
  pagination: Pagination;
  results: BookingHotelResult[];
}
// ...
export interface StarRating {
  tocLink: TocLink;
  caption: Caption;
  showAdditionalInfoIcon: boolean;
  symbol: string;
  value: number;
}
// ...
export interface BookingHotelResult {
  relocationMode: any;
  blocks: Block[];
  matchingUnitConfigurations: MatchingUnitConfigurations;
  showAdLabel: boolean;
  policies: Policies;
  persuasion: Persuasion;
  recommendedDatesLabel: any;
  soldOutInfo: SoldOutInfo;
  acceptsWalletCredit: boolean;
  inferredLocationScore: number;
  customBadges: CustomBadges;
  showPrivateHostMessage: boolean;
  bundleRatesAvailable: boolean;
  displayName: DisplayName;
  trackOnView: any[];
  propertySustainability: PropertySustainability;
  nbWishlists: number;
  mealPlanIncluded?: MealPlanIncluded;
  geniusInfo: any;
  isNewlyOpened: boolean;
  bookerExperienceContentUIComponentProps: any[];
  seoThemes: any[];
  description: any;
  isTpiExclusiveProperty: boolean;
  topPhotos: any[];
  priceDisplayInfoIrene: PriceDisplayInfoIrene;
  visibilityBoosterEnabled?: boolean;
  recommendedDate: RecommendedDate;
  ribbon?: Ribbon;
  hostTraderLabel: string;
  basicPropertyData: BookingBasicPropertyData;
  __typename: string;
  licenseDetails: any;
  location: HotelLocation;
  badges: any[];
  showGeniusLoginMessage: boolean;
}

export interface Block {
  hasCrib: boolean;
  __typename: string;
  blockMatchTags: BlockMatchTags;
  blockId: BlockId;
  freeCancellationUntil?: string;
  originalPrice: OriginalPrice;
  thirdPartyInventoryContext: ThirdPartyInventoryContext;
  onlyXLeftMessage?: OnlyXleftMessage;
  finalPrice: FinalPrice;
}

export interface BlockMatchTags {
  childStaysForFree: boolean;
  __typename: string;
}

export interface BlockId {
  packageId: string;
  policyGroupId: string;
  occupancy: number;
  mealPlanId: number;
  __typename: string;
  roomId: string;
}

export interface OriginalPrice {
  __typename: string;
  amount: number;
  currency: string;
}

export interface ThirdPartyInventoryContext {
  isTpiBlock: boolean;
  __typename: string;
}

export interface OnlyXleftMessage {
  tag: any;
  __typename: string;
  variables: any[];
  translation: string;
}

export interface FinalPrice {
  __typename: string;
  amount: number;
  currency: string;
}

export interface MatchingUnitConfigurations {
  commonConfiguration: CommonConfiguration;
  unitConfigurations: UnitConfiguration[];
  __typename: string;
}

export interface CommonConfiguration {
  nbAllBeds: number;
  nbBedrooms: number;
  bedConfigurations: BedConfiguration[];
  unitTypeNames: UnitTypeName[];
  unitId: number;
  nbBathrooms: number;
  name: any;
  localizedArea?: LocalizedArea;
  __typename: string;
  nbKitchens: number;
  nbUnits: number;
  nbLivingrooms: number;
}

export interface BedConfiguration {
  beds: Bed[];
  __typename: string;
  nbAllBeds: number;
}

export interface Bed {
  count: number;
  __typename: string;
  type: number;
}

export interface UnitTypeName {
  __typename: string;
  translation: string;
}

export interface LocalizedArea {
  unit: string;
  localizedArea: string;
  __typename: string;
}

export interface UnitConfiguration {
  nbBathrooms: number;
  name: string;
  __typename: string;
  nbLivingrooms: number;
  unitTypeId: number;
  unitId: number;
  localizedArea?: LocalizedArea2;
  nbKitchens: number;
  nbUnits: number;
  apartmentRooms: ApartmentRoom[];
  nbAllBeds: number;
  nbBedrooms: number;
  bedConfigurations: BedConfiguration2[];
  unitTypeNames: UnitTypeName2[];
}
// ...
export interface Policies {
  enableJapaneseUsersSpecialCase: any;
  __typename: string;
  showNoPrepayment: boolean;
  showFreeCancellation: boolean;
}

export interface Persuasion {
  nativeAdsCpc: any;
  geniusRateAvailable: boolean;
  preferredPlus: boolean;
  nativeAdsTracking: any;
  __typename: string;
  highlighted: boolean;
  bookedXTimesMessage: any;
  preferred: boolean;
  showNativeAdLabel: boolean;
  autoextended: boolean;
  sponsoredAdsData: any;
  nativeAdId: any;
}

export interface SoldOutInfo {
  __typename: string;
  alternativeDatesMessages: any[];
  messages: any[];
  isSoldOut: boolean;
}

export interface CustomBadges {
  showOnlineCheckinBadge: any;
  __typename: string;
  showParkAndFly: boolean;
  showIsWorkFriendly: boolean;
  showBhTravelCreditBadge: boolean;
  showSkiToDoor: boolean;
}

export interface DisplayName {
  __typename: string;
  text: string;
  translationTag: TranslationTag6;
}

export interface TranslationTag6 {
  translation: any;
  __typename: string;
}

export interface PropertySustainability {
  tier: Tier;
  isSustainable: boolean;
  facilities?: Facility[];
  certifications: any;
  levelId: string;
  chainProgrammes: any;
  __typename: string;
}

export interface Tier {
  __typename: string;
  type: string;
}

export interface Facility {
  id: number;
  __typename: string;
}

export interface MealPlanIncluded {
  mealPlanType: string;
  __typename: string;
  text: string;
}

export interface PriceDisplayInfoIrene {
  excludedCharges: ExcludedCharges;
  __typename: string;
  displayPrice: DisplayPrice;
  chargesInfo: ChargesInfo;
  priceBeforeDiscount: PriceBeforeDiscount;
  badges: Badge[];
  discounts: Discount[];
  useRoundedAmount: boolean;
  rewards: Rewards;
  taxExceptions: any[];
}

export interface ExcludedCharges {
  excludeChargesList: ExcludeChargesList[];
  excludeChargesAggregated: ExcludeChargesAggregated;
  __typename: string;
}
// ...
export interface DisplayPrice {
  amountPerStay: AmountPerStay;
  copy: Copy;
}

export interface ChargesInfo {
  __typename: string;
  translation: string;
}

export interface PriceBeforeDiscount {
  amountPerStay: AmountPerStay;
  __typename: string;
  copy: Copy;
}

export interface Badge {
  identifier: string;
  name: Name;
  tooltip: Tooltip;
  style: string;
  __typename: string;
}

export interface Tooltip {
  translation: string;
  __typename: string;
}

export interface Discount {
  itemType: string;
  name: Name;
  __typename: string;
  description: Description;
  amount: Amount;
  productId: string;
}

export interface Description {
  translation: string;
  __typename: string;
}

export interface Amount {
  currency: string;
  amount: string;
  __typename: string;
  amountUnformatted: number;
  amountRounded: string;
}

export interface Rewards {
  rewardsAggregated: RewardsAggregated;
  rewardsList: any[];
  __typename: string;
}

export interface RewardsAggregated {
  copy: Copy;
  __typename: string;
  amountPerStay: AmountPerStay;
}

export interface RecommendedDate {
  checkin: string;
  lengthOfStay: number;
  __typename: string;
  checkout: string;
}

export interface Ribbon {
  __typename: string;
  ribbonType: string;
  text: string;
}

export interface BookingBasicPropertyData {
  __typename: string;
  photos: BookingHotelPhotos;
  location: HotelLocation;
  accommodationTypeId: number;
  starRating?: StarRating;
  pageName: string;
  isTestProperty: boolean;
  isClosed: boolean;
  ufi: number;
  id: number;
  reviewScore: ReviewScore;
  externalReviewScore?: ExternalReviewScore;
  alternativeExternalReviewsScore: any;
}

export interface BookingHotelPhotos {
  __typename: string;
  main: Main;
}
// ...
export interface ReviewScore {
  reviewCount: number;
  score: number;
  showSecondaryScore: boolean;
  secondaryScore: number;
  __typename: string;
  showScore: boolean;
  secondaryTextTag: SecondaryTextTag;
  totalScoreTextTag: TotalScoreTextTag;
}

export interface SecondaryTextTag {
  translation?: string;
  __typename: string;
}

export interface TotalScoreTextTag {
  translation: string;
  __typename: string;
}

export interface ExternalReviewScore {
  __typename: string;
  totalScoreTextTag: TotalScoreTextTag;
  showScore: boolean;
  reviewCount: number;
  score: number;
}

export interface HotelLocation {
  displayLocation: string;
  beachWalkingTime: any;
  mainDistance: string;
  skiLiftDistance: any;
  publicTransportDistanceDescription: any;
  __typename: string;
  geoDistanceMeters: any;
  beachDistance: any;
  nearbyBeachNames: any[];
  countryCode: any;
}